lec5/requirements.txt
streamlit
google-generative-ai
python-dotenv
langchain
pyPDF2
chromadb